requirements.txt
telethon==1.32.1
schedule==1.2.0
python-dotenv==1.0.0 